Export named charts
