+++ packages/clay-charts-react/package.json
@@ -13,9 +13,8 @@
 	"license": "BSD",
 	"dependencies": {
 		"clay-charts-shared": "^1.0.0",
-		"react-billboardjs": "^1.4.3",
-		"react": "^16.5.2",
-		"react-dom": "^16.6.0"
+		"fbjs": "^1.0.0",
+		"react-billboardjs": "^1.4.3"
 	},
 	"peerDependencies": {
 		"react": "^16.5.2",
@@ -31,7 +30,7 @@
 		"css-loader": "^1.0.0",
 		"html-webpack-plugin": "^3.2.0",
 		"ncp": "^2.0.0",
-		"react": "^16.5.2",
+		"react": "^16.6.0",
 		"react-billboardjs": "^1.4.3",
 		"react-dom": "^16.6.0",
 		"style-loader": "^0.23.1",
